[Fix #155] Provide `assert_offense` and `assert_no_offenses` for custom cop development
Resolves #155.

This PR provides `assert_offense`, `assert_correction`, and `assert_no_offenses`
testing APIs for custom Minitest cop development.

Developers can use these APIs by requesting the following in custom Minitest cop test code:

```ruby
require 'rubocop/minitest/support'
```
