+++ lib/rubocop/minitest/support.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+# Require this file to load code that supports testing using Minitest.
+
+require 'rubocop'
+require 'minitest/autorun'
+require 'minitest/pride'
+require_relative 'assert_offense'
+
+Minitest::Test.include RuboCop::Minitest::AssertOffense
