+++ test/test_helper.rb
@@ -1,9 +1,6 @@
 # frozen_string_literal: true
 
-require 'rubocop'
 require 'rubocop-minitest'
-require 'minitest/autorun'
-require 'minitest/pride'
-require 'assertion_helper'
 
-Minitest::Test.include AssertionHelper
+# Require supporting files exposed for testing.
+require 'rubocop/minitest/support'
